+++ packages/domain/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./confidence";
 export * from "./gtin";
+export * from "./material-code";
 export * from "./models";